Add a test case for a bug that causes an assertion failure in the
Estimated hours taken: 6
Branches: main, release

Add a test case for a bug that causes an assertion failure in the
compiler.  It seems to be caused by the interaction of higher-order and
type specialization; a unification involving the construction of a
type_ctor_info (for the type being specialized) in a predicate created
by higher-order specialization seems to have an incorrect entry in the
instmap - this results in simplify incorrectly optimizing away the
unification, even though its output is need later.  (The unification is
the one involving V_20 in the predicate foldl_ho1/5).

This is the same bug that causes the module compiler/mode_robdd.tfeirn
to have to be compiled at a reduced optimization level.  We don't
currently have a fix for this (I'm still working on that), however a
workaround is to disable higher-order specialization for any affected
modules.  (In what is probably another bug, you don't seem to be able to
disable type-specialization, if a predicate has a type_spec pragma on
it).

tests/valid/Mmakefile:
tests/valid/Mercury.options:
tests/valid/ho_and_type_spec_bug.m:
	A test case that exposes a bug in some of the optimization
	passes.