Fix the test case tests/tabling/memo_non which was failing at the higher
Estimated hours taken: 0.5
Branches: main

Fix the test case tests/tabling/memo_non which was failing at the higher
levels of optimization, e.g. -O5, because the compiler was pushing the
call to marker/4 into the 'then' branch of the if-then-else (which is
the only place that actually uses the output of marker/4).  The reason
that this has started happening is that I recently modified the compiler
to take account of the default termination/exception properties of
foreign procs during optimization passes and marker/4 fulfills all the
criteria for something that the compiler thinks it can reorder
The fix is to make marker/4 impure.

tests/tabling/memo_non.m:
	Make marker/4 impure, so that the compiler doesn't
	reorder it at higher optimization levels.