Workaround a problem in the demangler so that mprof works when the
Estimated hours taken: 8
Branches: main, release

Workaround a problem in the demangler so that mprof works when the
library is compiled at -O5 and greater.  The problem is that the names
of the predicates generated by unused argument analysis were changed
(when Simon integrated unused argument analysis into the intermodule
analysis framework), but the code in the demangler was never updated
accordingly.  (The log message says that there was a change to the
demangler being planned at the same time, which is why it was never
updated - evidently it never happened.)

The above situation sometimes results in two different mangled names
demangling to the same name.  The workaround, for the moment, is to
disable all demangling, bar the fixing of mangled ASCII codes, for
predicates created by unused argument analysis.

profiler/demangle.m:
util/mdemangle.c:
	Add a workaround to the demanglers so that the profiler
	does not abort when processing procedures produced by
	unused argument analysis.