Add a new compiler option --force-disable-tracing which forces tracing to be
Estimated hours taken: 1
Branches: main

Add a new compiler option --force-disable-tracing which forces tracing to be
turned off even if compiling in a .debug or .decldebug grade.
This is necessary so that the declarative debugger can be compiled with no
tracing even in .debug or .decldebug grades.
This allows the declarative debugger to use tail recursion, improves its
performance and reduces the size of the installed browser library for
.debug and .decldebug grades.

This diff does not use the --force-disable-tracing option.  It is only added to
the compiler.  A subsequenct diff will use the option to turn off tracing by
default in the browser directory.

The declarative debugger will still call traced code in the library,
however, so this is not a general fix.
A more general solution would be to link in a non-traced version of the
library when compiling the declarative debugger, but that can be done later.

compiler/handle_options.m:
compiler/options.m:
	Add and handle the new option.  It overrides all other tracing
	options.