Allow constant propagation to be run in deep profiling grades.
Estimated hours taken: 0.5
Branches: main, release

Allow constant propagation to be run in deep profiling grades.
There's no reason not to run it, and doing so eliminates irrelevant
information from profiles, e.g appending of constant strings etc.

(We possibly use to allow this anyway, but the changes to the
way inlining is handled in the compiler disabled it; on the other
hand the deep profiling grades were broken for quite a while so
it's also possible we did not.)

compiler/handle_options.m:
	Allow constant propagation in deep profiling grades.