Add a new mode of operation, '--verify-profile', to mdprof_test.
Estimated hours taken: 0.5
Branches: main

Add a new mode of operation, '--verify-profile', to mdprof_test.
When run in this mode, mdprof_test reads in a deep profile and
tests that it is well-formed by running the usual startup
procedure on it.  It does not create a server process.

The intention is to use this in the test suite when testing
the .profdeep grades, e.g to prevent the situation where a
test case is passing but the Deep.data file contains rubbish.

deep_profiler/mdprof_test.m:
	Add an option '--verify-profile' that causes to
	mdprof_test to run in verification mode.