Fix some problems caught by the nightly tests and reorder part of the
Estimated hours taken: 1
Branches: main, release

Fix some problems caught by the nightly tests and reorder part of the
reference manual.

tabling/specified was failing because the compiler was attempting to perform
unused argument optimization on a tabled predicate.  The fix is to
disable unused argument optimization for all predicates that do not have an
evaluation method of eval_normal.

debugger/interactive was failing because of the recent change to how the
superhomogenous form of lambda expressions is constructed.  The fix is to
update the expected output.  This diff also removes an older expected output
for this test case that is no longer applicable.

Reorder the C specific section on the FLI pragmas in the reference
manual so that it matches that of the other languages.

compiler/unused_args.m:
	Do not optimize unused arguments for tabled predicates.

doc/reference_manual.texi:
	Rearrange the ordering of the foreign pragmas for C so that it matches
	that of the other languages.

tests/debugger/interactive.exp:
	Update the expected output for this test case.  The change in expected
	output is due to the recent change in how the compiler processes lambda
	expressions.

tests/debugger/interactive.exp2:
	Remove the contents of this file and replace them with a message
	saying that the file is currently unused.  This is expected output is
	out-of-date with respect to the .inp file.  (This version of the
	expected output was originally added because the compiler generated
	different output at -O1, however it no longer does that.)