Significantly improve the capabilities of the LLDS optimization that tries
Estimated hours taken: 20
Branches: main

Significantly improve the capabilities of the LLDS optimization that tries
to delay the creation of the stack frame, in the hope that on some computation
paths the frame won't need to be created at all. Previously, the delayed
setup of the stack frame could take place only when a block without a stack
frame fell through to a block that needed a stack frame. If block B1 jumped
to another block B2 that needed a frame, this was taken as meaning that B1
also had to have a frame. This was a problem, because if B1 ends with a
computed goto, some of whose targets need stack frames and some do not,
this limitation effectively gave all of them a stack frame, whether they
wanted it or not, and thus required them to execute the stack frame teardown
code.

This diff removes the limitation, optimization allows B1 in this case to not
have a stack frame. Instead of jumping to B2, B1 will not jump to a label B3
it inserts immediately before B2, the code at B3 setting up the stack frame
and falling through to B2. (We also insert code to jump around B3 if the
code immediately preceding it could fall into it accidentally.)

The new code in frameopt is conceptually cleaner than it was before, because
we now handle transitions from blocks that don't have a stack stack to blocks
that do in a much more uniform manner.

Most of the changes to other modules are to make the change to frameopt.m
easier to debug.

The motivation for the change was that we were beaten by YAP (Yet Another
Prolog) on the deriv benchmark due to the limitation of frameopt. I haven't
measured against YAP yet, but the runtime for 1.5 million iterations has been
reduced from about 20 seconds to about 13.

Since the compiler doesn't have any predicates that are both frequently used
and can benefit from the removal of that old limitation (which is why the
limitation wasn't really noticed before), there is no measurable effect
on the speed of the compiler itself.

compiler/frameopt.m:
	Effectively rewrite the optimization that delays stack frame creation
	along the lines above. The code for the optimization that keeps the
	stack frame for recursive calls if possible is unaffected.

	If the new option --frameopt-comments is specified, insert into the
	generated LLDS code a nicely formatted description of the main
	frameopt.m data structures. These are much easier to read that the
	term browser in the debugger.

compiler/options.m:
	Add the new developer-only option --frameopt-comments.

compiler/llds_out.m:
	Change the way we output comments to make the coments generated by
	frameopt.m easier to read. (We output comments only if --auto-comments
	is given, which it usually isn't.)

compiler/opt_debug.m:
	Provide the functionality of printing local labels in an easier-to-read
	form that doesn't repeat the (possibly long) procedure name. Local
	labels can now be printed as e.g. local_15.

	Rewrite the module to use functions instead of predicates for appending
	strings, since this makes the code shorter, easier to use and to read.
	The original code was written before Mercury had functions.

compiler/switch_util.m:
	When gathering information about switches, return the cons_id with each
	goal.

	Switch to four-space indentation.

compiler/tag_switch.m:
	When generating code for switches, insert a comment at the start of
	each case saying what cons_id it is for, using the new information from
	switch_util. This is to make the generated code easier to understand.

	Switch to four-space indentation.

compiler/ml_tag_switch.m:
	Conform to the change in switch_util.

compiler/optimize.m:
	Conform to the slightly modified interface of frameopt.m.

	Switch to four-space indentation.

compiler/peephole.m:
	Switch to four-space indentation, and fix some coding style issues.

compiler/basic_block.m:
	When dividing a procedure body into basic blocks, remember for each
	block whether it could be fallen into. This modification is not
	strictly required for this change, since frameopt has its own
	(specialized) code for creating basic blocks, but it could be useful
	in the future.

compiler/dupelim.m:
compiler/use_local_vars.m:
	Conform to the change in basic_block.m.