Make the default sizes of memory zones proportional to the word size.
Estimated hours taken: 1
Branches: main

Make the default sizes of memory zones proportional to the word size.  None
of the default sizes change on a 32-bit architecture.

The rationale for this is that the amount of memory used by a Mercury program
for a given input is generally proportional to the word size.  Therefore,
before this change, a program that works on a 32-bit machine may overflow one
of the memory zones on a 64-bit machine unless special precautions are taken.
(In particular, the compiler cannot be bootchecked in asm_fast.gc.tr.debug
on a 64-bit machine without such a change.)

Add variants of size-setting options in the MERCURY_OPTIONS environment
variable that take kilo-words as the units.

runtime/mercury_wrapper.c:
	Modify the defaults and parse the new options.

	Add a comment that says the user guide should be updated if the
	defaults change.

doc/user_guide.texi:
	Update the documentation corresponding to this.  Also, fix some
	out-of-date numbers.