My previous change to support --generate-dependency-file was incorrect
Estimated hours taken: 0.5
Branches: main

My previous change to support --generate-dependency-file was incorrect
as it required all the .int files to be available before being able to
build the '.d' file.

Instead move the generation of the '.d' file to the same place as the
generation of the '.dep' and '.dv' file.

compiler/mercury_compile.m:
	Call generate_file_dependency_file and generate_module_dependency_file.

compiler/modules.m:
	Add generate_file_dependency_file and generate_module_dependency_file.
	This is done by refactoring  generate_dependencies so that it can be
	used to generate the '.d' file only.
	Remove my previous change to generate the '.d' file in mercury_compile.